backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
openai>=1.50.0
httpx>=0.27.0,<0.29.0
pdfplumber==0.11.0
tiktoken==0.7.0
langdetect==1.0.9
fpdf2==2.7.9
python-multipart==0.0.9
pgvector==0.2.5
